+++ src/QueryBuilder/AbstractDQLQueryBuilder.php
@@ -256,28 +256,17 @@ abstract class AbstractDQLQueryBuilder implements DQLQueryBuilderInterface
             return '';
         }
 
-        /**
-         * @psalm-var array<
-         *   array-key,
-         *   array{
-         *     0?:string,
-         *     1?:array<array-key, Query|string>|string,
-         *     2?:array|ExpressionInterface|string|null
-         *   }|null
-         * > $joins
-         */
         foreach ($joins as $i => $join) {
+            /** @psalm-suppress DocblockTypeContradiction */
             if (!is_array($join) || !isset($join[0], $join[1])) {
                 throw new Exception(
-                    'A join clause must be specified as an array of join type, join table, and optionally join '
-                    . 'condition.'
+                    'A join clause must be specified as an array of join type, join table, and optionally join condition.',
                 );
             }
 
-            /* 0:join type, 1:join table, 2:on-condition (optional) */
             [$joinType, $table] = $join;
 
-            $tables = $this->quoteTableNames((array) $table, $params);
+            $tables = $this->quoteTableNames(is_array($table) ? $table : [$table], $params);
 
             /** @var string $table */
             $table = reset($tables);
@@ -644,9 +633,11 @@ abstract class AbstractDQLQueryBuilder implements DQLQueryBuilderInterface
                     $table = $this->quoter->quoteTableName($table);
                 }
                 $tables[$i] = "$table " . $this->quoter->quoteTableName($i);
-            } elseif ($table instanceof ExpressionInterface && is_string($i)) {
+            } elseif ($table instanceof ExpressionInterface) {
                 $table = $this->buildExpression($table, $params);
-                $tables[$i] = "$table " . $this->quoter->quoteTableName($i);
+                $tables[$i] = is_string($i)
+                    ? "$table " . $this->quoter->quoteTableName($i)
+                    : $table;
             } elseif (is_string($table) && !str_contains($table, '(')) {
                 $tableWithAlias = $this->extractAlias($table);
                 if (is_array($tableWithAlias)) { // with alias